Introduce helper_for_tests header again
