+++ tests/mock_data/test_gamestates.txt
@@ -4,4 +4,4 @@ position startpos moves e2e4 c7c5 g1f3 b8c6 f1b5 e7e6 e1g1 g8e7 b5a4 a7a6 c2c4 d
 position fen 8/6pk/5pn1/1q5p/7P/3P2P1/4QP2/4N1K1 b - - 0 1
 position fen 8/6pk/5pn1/1q5p/7P/3P2P1/4QP2/4N1K1 b - - 0 1 moves h7g8 e1g2 g8f8 g1h2 g6e5 g2f4
 position startpos moves e2e4 e7e6 d2d4 d7d5 b1d2 c7c5 e4d5 e6d5 g1f3 g8f6 f1b5 c8d7 b5d7 b8d7 e1g1 f8e7 d4c5 d7c5 d2b3 c5e4 f3d4 e8g8 c2c3 f8e8 c1e3 g7g6 f1e1 d8c7 d1f3 e7d6 g2g3 f6d7 f3d1 a7a6 d4c2 a8d8 f2f3 e4f6 d1e2 h7h5 e2f2 e8e5 g1g2 d8e8 e3f4 e5e1 f4d6 c7d6 a1e1 e8e1 c2e1 d7e5 b3d4 f6d7 f2e2 d7c5 e1c2 e5c6 d4c6 b7c6 c2d4 g8f8 b2b4 c5e6 d4e6 d6e6 e2e6 f7e6 f3f4 f8e7 g2f3 e7d6 f3e3 c6c5 e3d3 c5b4 c3b4 d6c6 a2a4 c6d6 d3d4 d6d7 d4c3 d7d6 c3b3 d6c6 h2h3 c6c7 b3c3 c7d7 c3d3 d7c7 d3d4 c7d6 d4e3 d6c6 g3g4 h5g4 h3g4 c6d6 e3e2 d6c7 e2d1 c7d7 d1c1 d7d6 c1c2 d6c7 c2c3 c7d6 c3d3 d6c7 d3c2 c7d6 c2c3 d6d7 c3d4 d7d6 g4g5 d6c6 d4c3 c6c7 c3d4 c7c6 d4e5 c6b7 e5e6 d5d4 f4f5 g6f5 g5g6 d4d3 g6g7 d3d2 g7g8q d2d1q g8f7 b7b6 a4a5 b6b5 f7b7 b5a4 b7a6 d1b3 e6f6 b3c3 f6f5 c3b4 f5e6 b4c5 a6b6 c5b5 b6b5 a4b5 e6d5 b5a5
-8/8/8/k2K4/8/8/8/8 w - - 0 79
+position fen 8/8/8/k2K4/8/8/8/8 w - - 0 79
